+++ noodles-sam/src/async.rs
@@ -1,3 +1,4 @@
 //! Async SAM.
 
+pub mod fs;
 pub mod io;